requirements.txt
Django==4.2.7
django-bootstrap5==23.3
django-bootstrap-modal-forms
django-bootstrap-datepicker-plus==5.0.4